+++ sbuildr/graph/node.py
@@ -1,15 +1,16 @@
 from sbuildr.tools import compiler, linker
 from sbuildr.tools.flags import BuildFlags
 from sbuildr.logger import G_LOGGER
+from sbuildr.misc import paths
 
 from typing import List
 import os
 
 # Represents a node in a dependency graph that tracks a path on the filesystem.
 class Node(object):
-    def __init__(self, path: str, inputs: List["Node"]=[], name=""):
+    def __init__(self, path: str, inputs: List["Node"]=[]):
         self.path = path
-        self.name = name or os.path.basename(path)
+        self.basename = os.path.basename(path)
         self.inputs: List[Node] = []
         self.outputs: List[Node] = []
         G_LOGGER.debug(f"Constructing {self} with {len(inputs)} inputs: {inputs}")
@@ -17,7 +18,7 @@ class Node(object):
             self.add_input(inp)
 
     def __str__(self):
-        return f"{self.name}: {self.path}"
+        return f"{self.basename}: {self.path}"
 
     def __repr__(self):
         return f"{self}: {self.path} (at {hex(id(self))})"
@@ -25,7 +26,7 @@ class Node(object):
     # Returns a string representation of the dependency graph for this node.
     def dependency_graph_str(self, tab_depth=0):
         tab = '\t'
-        out = f"{tab * tab_depth}{self.name}\n"
+        out = f"{tab * tab_depth}{self.basename}\n"
         for inp in self.inputs:
             out += f"{inp.dependency_graph_str(tab_depth + 1)}\n"
         return out
@@ -36,15 +37,15 @@ class Node(object):
         self.inputs.append(node)
 
 class SourceNode(Node):
-    def __init__(self, path: str, inputs: List["SourceNode"]=[], include_dirs: List[str]=None, name=""):
-        super().__init__(path, inputs, name)
+    def __init__(self, path: str, inputs: List["SourceNode"]=[], include_dirs: List[str]=None):
+        super().__init__(path, inputs)
         # All include directories required for this file.
         self.include_dirs = include_dirs
 
 class CompiledNode(Node):
     # These include_dirs are user-specified, since any scanned dirs would be in the SourceNode.
-    def __init__(self, path: str, input: SourceNode, compiler: compiler.Compiler, include_dirs: List[str]=[], flags: BuildFlags=BuildFlags(), name=""):
-        super().__init__(path, [input], name)
+    def __init__(self, path: str, input: SourceNode, compiler: compiler.Compiler, include_dirs: List[str]=[], flags: BuildFlags=BuildFlags()):
+        super().__init__(path, [input])
         self.compiler = compiler
         # All include directories required for this file.
         self.include_dirs = include_dirs
@@ -55,10 +56,34 @@ class CompiledNode(Node):
             G_LOGGER.critical(f"Cannot create a CompiledNode with more than one source. This node already has one input: {self.inputs}")
         super().add_input(node)
 
+class Library(Node):
+    # TODO: Add search_dirs parameter?
+    def __init__(self, name: str=None, path: str=None, ld_dirs: List[str]=[]):
+        """
+        Represents a library.
+
+        :param name: The name of the library.
+        :param path: A path to the library.
+        :param ld_dirs: A list of directories required for loading this library. This would generally include directories containing libraries that this library is linked against. For example, if the project requires ``liba``, and ``liba`` is linked against ``libb``, then ``ld_dirs`` should include the containing directory of ``libb``.
+
+        Note that either a name or path must be provided. If a name is provided, then the containing directory for this library should be provided to ``ld_dirs``, unless it is in the default linker/loader search path.
+        """
+        if not (name or path):
+            G_LOGGER.critical(f"Either a name or path must be provided to find a library")
+
+        self.path = path
+        # TODO: FIXME: This will not handle non-standard library names (e.g. not in the form lib<name>.so)
+        self.name = name or paths.libname_to_name(os.path.basename(path))
+        self.basename = self.name or self.path
+        self.ld_dirs = [os.path.abspath(dir) for dir in ld_dirs]
+        if self.path:
+            self.ld_dirs.insert(0, os.path.dirname(self.path))
+
 # Only CompiledNodes in the inputs list are passed on to the linker.
-class LinkedNode(Node):
-    def __init__(self, path: str, inputs: List[Node], linker: linker.Linker, libs: List[str]=[], lib_dirs: List[str]=[], flags: BuildFlags=BuildFlags(), name=""):
-        super().__init__(path, inputs, name)
+class LinkedNode(Library):
+    def __init__(self, path: str, inputs: List[Node], linker: linker.Linker, libs: List[str]=[], lib_dirs: List[str]=[], flags: BuildFlags=BuildFlags()):
+        super().__init__(path=path, ld_dirs=lib_dirs)
+        Node.__init__(self, path, inputs)
         self.linker = linker
         self.libs = libs
         self.lib_dirs = lib_dirs